// src/generated/models/patient-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records - Appointment scheduling and management - Vitals recording and examination - Prescription and investigation management - IPD/OPD workflow 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BloodGroup } from './blood-group';
// May contain unused imports in some cases
// @ts-ignore
import type { EmergencyContactResponse } from './emergency-contact-response';
// May contain unused imports in some cases
// @ts-ignore
import type { Gender } from './gender';
// May contain unused imports in some cases
// @ts-ignore
import type { InsuranceResponse } from './insurance-response';
// May contain unused imports in some cases
// @ts-ignore
import type { MedicalHistoryResponse } from './medical-history-response';
// May contain unused imports in some cases
// @ts-ignore
import type { PatientStatus } from './patient-status';
// May contain unused imports in some cases
// @ts-ignore
import type { PreferredContactMethod } from './preferred-contact-method';

export interface PatientResponse {
    'id'?: number;
    'patientId'?: string;
    'firstName'?: string;
    'lastName'?: string;
    'dateOfBirth'?: string;
    'age'?: number;
    'gender'?: Gender;
    'bloodGroup'?: BloodGroup;
    'aadharNumber'?: string;
    'photoUrl'?: string;
    'mobileNumber'?: string;
    'emailId'?: string;
    'preferredContactMethod'?: PreferredContactMethod;
    'addressLine1'?: string;
    'city'?: string;
    'district'?: string;
    'state'?: string;
    'pinCode'?: string;
    'emergencyContacts'?: Array<EmergencyContactResponse>;
    'insurance'?: InsuranceResponse;
    'medicalHistory'?: MedicalHistoryResponse;
    'status'?: PatientStatus;
    'createdAt'?: string;
    'updatedAt'?: string;
}